Non interactive version.
The interactive version of the translator can be a little cumbersome if what
you want is to translate only a single word. From now on you can make your
translations passing the text as arguments, e.g.

    $ tl alibi
    verb:
    -----
    presentar una coartada

    noun:
    -----
    coartada, excusa

Or

    $ tl p clincher
    noun:
    -----
    punto clave

Or

    $ tl p And here is the clincher.
    Y aquí es el factor decisivo.

The concatenation of all arguments will be taken as the input text.