Create fake API server.
Our test suite depends heavily on API calls to external servers and we
don't want to be banned by those servers, right? To solve this problem I
have created a mini server application that will serve canned responses
to our program in the test phase.

To make this work we have to install dnsmasq and configure it to
redirect requests to the external sites to our local machine. I will
document this in our wiki

In order to run the server at the 80 port run the server with `rvmsudo`:
`rvmsudo fake_api_server.rb`.